Big changes
We are nearly done guys. Now the tasks are to create a settings editor, which opens upon a button press to Menu->Settings (The settings tab in the tab field will be deleted soon). A distribution editor would also be fancy - a window in which you can "draw" a packet distribution with your mouse and save it to disk. This distribution could then be loaded into the simulation. Furthermore we need to refine the assignment of packets to the cpu's as we did not yet implement this (hardest feature of them all). I thought, that we should look at the prediction and the current state and then set the clusters mode by setMode function to a more (or less) aggressive allocation algorithm. To ensure good results we do also have to define a couple of parameters which allow for flawless program execution (minimal packet loss and maximum efficiency). Even if our program is running without any errors - without balanced parameters (settings) we will hit a wall.

Have fun getting grilled by the sun and temperatures!